cli order to its own file
